Move the sentry user generation logic into the telemetry component.
